+++ packages/vsce/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "zowe-native-proto-vsce" extension will be documented
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Recent Changes
+
+- Added unit tests for `SshConfigUtils.ts` file. [#586] (https://github.com/zowe/zowe-native-proto/pull/586)
+
 ## `0.1.10`
 
 - Updated the `Zowe-SSH: Connect to Host...` command to prompt the user to choose a deploy directory.[#527] (https://github.com/zowe/zowe-native-proto/issues/527)